The message processing process has changed: More sign-in handlers do not accept the two message request and response.  Handlers to the input accepted in the context and formation of the answer is cites logic users of handlers.#7

+++ libactor_zeta_io/actor-zeta/network/connection_identifying.hpp
@@ -9,7 +9,7 @@ namespace actor_zeta {
 /// @enum type_connect
 /// @brief
 ///
-        enum class type_connect : int {
+        enum class type_connect : uint8_t {
             tcp,
             udp
         };
